+++ spec/integration/bills_kitchen_spec.rb
@@ -6,11 +6,11 @@ describe "bills kitchen" do
   include Helpers
 
   describe "tools" do
-    it "installs Chef-DK 0.2.0" do
-      run_cmd("chef -v").should match('Chef Development Kit Version: 0.2.0')
+    it "installs ChefDK 0.3.5" do
+      run_cmd("chef -v").should match('Chef Development Kit Version: 0.3.5')
     end
-    it "installs Vagrant 1.6.3" do
-      run_cmd("vagrant -v").should match('1.6.3')
+    it "installs Vagrant 1.6.5" do
+      run_cmd("vagrant -v").should match('1.6.5')
     end
     it "installs ssh.exe" do
       run_cmd("ssh -V").should match('OpenSSH_6.0p1, OpenSSL 1.0.1c 10 May 2012')
@@ -19,7 +19,7 @@ describe "bills kitchen" do
       run_cmd("rsync --version").should match('rsync  version 3.0.9')
     end
     it "installs Git 1.9" do
-      run_cmd("git --version").should match('git version 1.9.0')
+      run_cmd("git --version").should match('git version 1.9.4')
     end
     it "installs kdiff3" do
       marker_file = "#{BUILD_DIR}/merged.md"
@@ -77,14 +77,23 @@ describe "bills kitchen" do
     end
 
     describe "chefdk ruby" do
-      it "installs Chef 11.14.0.rc.2" do
-        run_cmd("knife -v").should match('Chef: 11.14.0.rc.2')
+      it "installs Chef 11.18.0.rc.1" do
+        run_cmd("knife -v").should match('Chef: 11.18.0.rc.1')
       end
-      it "uses the chef-dk embedded gemdir" do
+      it "uses the ChefDK embedded gemdir" do
         run_cmd("#{CHEFDK_RUBY}/bin/gem environment gemdir").should match("#{CHEFDK_RUBY}/lib/ruby/gems/2.0.0")
       end
-      it "has 'bundler (1.5.2)' gem installed" do
-        gem_installed "bundler", "1.5.2"
+      it "does NOT install gems into $HOME/.chefdk" do
+        gem_env = run_cmd("#{CHEFDK_RUBY}/bin/gem environment")
+        gem_env.should match('"update" => "--no-user-install"')
+        gem_env.should match('"install" => "--no-user-install"')
+        Dir["#{CHEFDK_HOME}/gem/ruby/2.0.0/"].should be_empty
+      end
+      it "has ChefDK verified to work via `chef verify`" do
+        cmd_succeeds "chef verify"
+      end
+      it "has 'bundler (1.6.7)' gem installed" do
+        gem_installed "bundler", "1.6.7"
       end
       it "has 'knife-audit (0.2.0)' plugin installed" do
         knife_plugin_installed "knife-audit", "0.2.0"
@@ -104,11 +113,11 @@ describe "bills kitchen" do
       it "has 'vagrant-omnibus (1.4.1)' plugin installed" do
         vagrant_plugin_installed "vagrant-omnibus", "1.4.1"
       end
-      it "has 'vagrant-cachier (0.8.0)' plugin installed" do
-        vagrant_plugin_installed "vagrant-cachier", "0.8.0"
+      it "has 'vagrant-cachier (1.1.0)' plugin installed" do
+        vagrant_plugin_installed "vagrant-cachier", "1.1.0"
       end
-      it "has 'vagrant-berkshelf (3.0.0)' plugin installed" do
-        vagrant_plugin_installed "vagrant-berkshelf", "3.0.0"
+      it "has 'vagrant-berkshelf (3.0.1)' plugin installed" do
+        vagrant_plugin_installed "vagrant-berkshelf", "3.0.1"
       end
     end
   end